requirements.txt
anndata==0.10.8
array_api_compat==1.8
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.9.1
blosc2==2.7.1
bypy==1.8.5
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
factor_analyzer==0.3.2
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
h5py==3.11.0
huggingface-hub==0.24.5
idna==2.10
igraph==0.11.6
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
legacy-api-wrap==1.4
leidenalg==0.10.2
llvmlite==0.43.0
lxml==4.5.2
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.0.8
multiprocess==0.70.16
munkres==1.1.4
natsort==8.4.0
nbformat==5.5.0
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numexpr==2.10.0
numpy==1.26.3
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pandas-market-calendars==1.4.2
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.2.0
pip==24.2
pkgconfig==1.5.5
platformdirs==4.2.2
plotly==4.14.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
pyvips==2.2.3
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.24.0
requests-toolbelt==1.0.0
retrying==1.3.4
rpds-py==0.20.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==71.1.0
setuptools-scm==8.1.0
sinfo==0.3.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.13.1
tables==3.9.2
texttable==1.7.0
threadpoolctl==3.5.0
timm==1.0.8
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
unicodedata2==15.1.0
urllib3==1.25.11
wcwidth==0.2.13
wheel==0.43.0
